requirements.txt
aiobotocore==2.5.4
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
botocore==1.31.17
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
cryptography==41.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.4
django-allauth==0.54.0
django-autoslug==1.9.9
django-crispy-forms==2.0
django-environ==0.10.0
django-gunicorn==0.1.1
django-phonenumber-field==7.1.0
django-tinymce==3.6.1
frozenlist==1.4.0
fsspec==2023.9.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
lxml==4.9.3
multidict==6.0.4
mysqlclient==2.2.0
oauthlib==3.2.2
openpy==0.0.2
outcome==1.2.0
packaging==23.1
phonenumbers==8.13.19
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
PyPDF2==3.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3fs==2023.9.2
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
